+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Revision history for ocaml-export
 
+## 0.12.0.0 -- 2018-12-28
+* Fix Decode record case when NamedConstructor is an OCamlRef, not a primitive.
+
+## 0.11.0.0 -- 2018-12-22
+* Support record, encode, decode for type that has a type parameter filled with a non-primitive type (eg. newtype Z = Z (Maybe X))
+
 ## 0.10.0.0 -- 2018-12-07
 
 * Support sum-type constructors with non-primitive types